% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ActiveHighSWE}
\alias{ActiveHighSWE}
\title{Peak SWE data from 1958 to 2021}
\format{
A list of 58 data.frames
}
\usage{
data(ActiveHighSWE)
}
\description{
Contains only the peak SWE measurement for each year at stations active in 2021.
}
\keyword{datasets}
