% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{activeStations}
\alias{activeStations}
\title{Complete snowpack data from 1958 to 2021}
\format{
A list of 58 data.frames
}
\usage{
data(activeStations)
}
\description{
Contains all data at stations active in 2021.
}
\keyword{datasets}
