% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ECCC_param_codes}
\alias{ECCC_param_codes}
\title{Parameter names and codes for downloading ECCC water quality data}
\format{
A data.frame
}
\usage{
data(ECCC_param_codes)
}
\description{
Contains parameter codes with matching parameter names and classes.
}
\keyword{datasets}
