% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movingKendallSlope.R
\name{movingKendallSlope}
\alias{movingKendallSlope}
\title{Moving Kendall Slope on Single or Regional Time-Series}
\usage{
movingKendallSlope(
  data,
  valuesColumn,
  timeColumn,
  start,
  end,
  valuesUnits = "mm",
  timeUnits = "years",
  period = 10
)
}
\arguments{
\item{data}{A data.frame or a list containing the data.frames (one per site or season) to test. Must contain a column of type numeric with the data to be analyzed, and a matching integer column in sequence representing years, months, days, or other regular interval. Data frames in a list can be of different lengths but must have identical column names. Use as.data.frame if selecting a list element.}

\item{valuesColumn}{The column name of the values to be tested.}

\item{timeColumn}{The column name of the date/time representative integer values.}

\item{start}{The first data point (year, month, day, etc) that will be considered in the rolling trend calculations (first calculated slope is from start:(end + period)). This year must be present in at least one of the data.frames in the supplied list.}

\item{end}{The last data point (year, month, day, etc) considered in the rolling trend calculations.}

\item{valuesUnits}{A character vector describing the data you are analyzing. Defaults to "mm".}

\item{timeUnits}{A character vector describing the time units ("days", "years", "months").}

\item{period}{The length of the moving window. Defaults to 10.}
}
\value{
A data.frame with three columns: one column of p-values, one column of regional Kendall slopes (for lists) or Sen's slopes (data.frame), and one column showing the last data point (year, month, day, etc) used in the calculations.
}
\description{
Calculates a Mann-Kendall significance level and a Sen's slope (if single data.frame) or regional/seasonal Kendall slope (if list of data.frames) for a given time window within a larger time interval, with the time window moving incrementally to cover the specified time interval.
}
