% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyMK.R
\name{tidyMK}
\alias{tidyMK}
\title{Tidy Mann Kendal Result Table}
\usage{
tidyMK(ygDataList, valueColumn, idLabel)
}
\arguments{
\item{ygDataList}{A list of data.frame or tibble data objects providing station data}

\item{valueColumn}{The column name, character or integer, of the data values to be tested}

\item{idLabel}{A character string to label the id column}
}
\value{
A data.frame of test results: Mann-Kendall tau and 2-sided p-value. One row per dataframe in the provided list.
}
\description{
This function uses the \code{MannKendall()} function from the Kendall package and applies the default two-sided test.  The values tested must be either a vector of numeric or time series values.
}
\details{
Calculate and return tidy data.frame of Mann Kendall results
}
