% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidySens.R
\name{tidySens}
\alias{tidySens}
\title{Tidy Sens Slope Result Table}
\usage{
tidySens(data, Column, idLabel = "Site")
}
\arguments{
\item{data}{A data.frame or list of data.frame objects providing data. Data points must be ordered from oldest to youngest in each data.frame. Use as.data.frame if selecting an item from a list.}

\item{Column}{The column name, character or integer, of the data values to be tested.}

\item{idLabel}{A character string to label the id column (eg. "Station", "Site", "Creek").}
}
\value{
A data.frame of test results: number of points and high and low confidence intervals. One row per dataframe in the provided list.
}
\description{
Uses the \code{sens.slope()} function from the trend package to calculate a slope estimate with 95\% confidence intervals and a significance level (p-value) of the slope estimate. The values tested must be either a numeric vector or time series object, present as a column in a data.frame or list of data.frames.
}
\details{
Calculate and return a tidy data.frame of Sen's slope results
}
