% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsFetch.R
\name{tsFetch}
\alias{tsFetch}
\title{Utility for automatically fetching time-series data}
\usage{
tsFetch(
  TSlist,
  sources = "all",
  AQlogin = c("gtdelapl", "WQ*2021!"),
  HYlogin = NULL
)
}
\arguments{
\item{TSlist}{The data.table output of WRBTrends::stationMeta, containing at minimum the default columns specified by that function.}

\item{sources}{The list of sources you wish to download data for. "all" means all the sources listed in the "Data location" column of the input data.table. Can also specify from "Aquarius", "EQWin", "Snow Survey", "ECCC", or "Workbook" if you want to exclude any sources. If specifying "Workbook" all time-series must be in individual workbook tabs with columns "datetime" and "value".}

\item{AQlogin}{The login parameters for Aquarius in format c("username", "password"). Leave NULL if you are not fetching from Aquarius, in which case you should either ensure it is not specified in the input data.table or is excluded under the source parameter.}

\item{HYlogin}{The login parameters for HYDAT in format c("username", "password"). Leave NULL if you are not fetching from Aquarius, in which case you should either ensure it is not specified in the input data.table or is excluded under the source parameter.}
}
\value{
A list containing one element (tibble) per time-series, with processing performed to standardize the time-series to one common format. Designed to act as input to the WRBTrends:: ???? and ???? functions.
}
\description{
This script takes the data.table output  from the stationMeta function (truncated if you desire fewer stations) and fetches time-series information for the stations contained in the "Site name", "Location identifier" columns.
If you are fetching information from the EQWin or Snow Survey Access databases you MUST have access to the X drive on your machine. Aquarius data can be fetched from anywhere with an internet connection.
}
\details{
BE AWARE that this function may take a long time to execute, perform work on a multi-core machine with lots of memory and a good internet connection!
}
